+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Follows [Semantic Versioning](https://semver.org/).
 ### Added
 
 - Specify authors to save when creating the coAuthor file.
+- Clean up unused features in author file: `coAuthor`, `author` format functions and no need for the `write` method.
+- Convert `GitAuthors` function to TypeScript and define new internal type `CoAuthorSchema`.
 
 ## git-mob 3.1.1